refactor: fix typo in folder components
