+++ src/app/components/card/card.component.scss
