+++ src/app/pages/layout-components/layout/layout.component.ts
@@ -1,6 +1,6 @@
 import { Component, EventEmitter, Input, OnInit, Output } from '@angular/core';
 import { CommonModule } from '@angular/common';
-import { CardComponent } from 'src/app/comonents/card/card.component';
+import { CardComponent } from 'src/app/components/card/card.component';
 import { MatProgressSpinnerModule } from '@angular/material/progress-spinner';
 import { BreakpointObserver, Breakpoints, LayoutModule } from '@angular/cdk/layout';
 import { Observable, map, of } from 'rxjs';
